Fix cliff cutting
The cutting method was wrong and also the method to calculate the new
height after cutting did not work out.

We need to set a cut edge to the height corresponding to the average
height of still connected centroids.

This way we ensure that edges that are *not* cut fit height-wise and
form a stripe along a cliff side.

The colors in the visualization represent walkability.

green and yellow are walkable from the beach. grey and light-grey are
areas that are unreachable.