Refactor editor
We reorganize the way we supply tiles with ids and also how these ids
are arranged.

If a tile is larger than 1x1 it has to define an idCount definining how
many ids it needs and then provide an sizeX x sizeZ sized pattern

A 0 in that pattern is meant to be the location of the (oversized) tile.

So the house is now 3 x 3 and uses a pattern

    1,1,1,
    1,0,1,
    1,1,1,

ensuring that the central 0-id is always surrounded by special "empty"
tiles. Several houses can share those special empties then, but other
tiles stay away.