Refactor tiles / rules
The "special empty" around houses idea isn't working. It smears those
tiles all over the map often without creating an actual house. Not
surprising at 8 to 1 odds :\

This changes the pattern format to be able to use both relative local
ids and absolute ids.

Our house pattern now is

    0,0,0
    0,-1,0
    0,0,0

With -1 being the local id and -2 id +1 etc.

This also changes the editor handling a bit so that the cleanup function
is always invoked when changing values.

This way we can have 0 tiles as part of the pattern and remove our house
when someone changes that, but *not* if someone sets down a house right
next to ours which shares a row of 0s.